wpt_babel/content-security-policy/object-src/object-src-url-blocked.html

<!DOCTYPE html>
<html>
 <head>
  <meta content="object-src 'none'; script-src 'self' 'unsafe-inline';" http-equiv="Content-Security-Policy"/>
  <script src="/resources/testharness.js">
  </script>
  <script src="/resources/testharnessreport.js">
  </script>
 </head>
 <body>
  <script>
   "use strict";

var t = async_test("Should block the object and fire a spv");
var expected = 3;
var count = 0;
window.addEventListener('securitypolicyviolation', t.step_func(function (e) {
  count++;
  assert_equals(e.violatedDirective, "object-src");
  if (count == expected) {
    t.done();
  }
}));
function unexpectedObjectLoaded() {
  t.step(function () {
    assert_unreached('CSP should block this object from loading');
  });
}
  </script>
  <object data="/content-security-policy/support/fail.png" onload="unexpectedObjectLoaded()" type="image/png">
  </object>
  <object data="/content-security-policy/support/fail.png" onload="unexpectedObjectLoaded()">
  </object>
  <object data="application/x-webkit-test-netscape" onload="unexpectedObjectLoaded()">
  </object>
 </body>
</html>
